requirements.txt
langchain==1.2.12
langgraph==1.1.1
redis[hiredis]==7.3.0
langchain-openai==1.1.11
fastapi==0.135.1
uvicorn==0.41.0
aio-pika==9.6.1
qdrant-client>=1.10.0
python-dotenv==1.2.2
pydantic-settings==2.13.1
numpy
backoff==2.2.1
scikit-learn
PyMuPDF==1.27.2
python-docx==1.2.0
markdown==3.10.2
tiktoken==0.12.0
requests==2.32.3
httpx
fastembed
openpyxl
pandas